appends text at the end of file

+++ 0x15-file_io/2-append_text_to_file.c
@@ -29,7 +29,7 @@ int append_text_to_file(const char *filename, char *text_content)
 	{
 		text_len = strlen(text_content);
 		b_write = write(fileno(file), text_content, text_len);
-		if (b_write != text_len)
+		if (b_write != (ssize_t)text_len)
 		{
 		fclose(file);
 		return (-1);
